moved helper functions to utils/math
